Create vendor-item-card component
