+++ src/components/VendorDashboard/VendorDashboard.css
@@ -4,7 +4,7 @@
   background-size: cover;
 
   height: 100vh;
-  font-family: "Inter";
+  font-family: 'Inter';
   color: black;
   gap: 4rem;
   display: flex;
@@ -96,81 +96,11 @@
   background-color: #343a40;
 }
 
-.vendor-item-card {
-  display: flex;
-  margin: 2rem;
-  padding: 1.5rem;
-  border-radius: 5px;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.3);
-  background-color: #e5e7ebe2;
-}
-
-.vendor-item-image img {
-  max-width: 12rem;
-  max-height: 12rem;
-  object-fit: cover;
-  border-radius: 5;
-  margin-right: 1rem;
-}
-
-.item-details {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 1rem;
-}
-
-.item-input {
-  width: 100%;
-  padding: 5px;
-  margin: 5px 0;
-  border: none;
-}
-
-.item-btns {
-  margin-top: 10px;
-}
-
-.save-changes,
-.edit-changes {
-  padding: 8px 15px;
-  cursor: pointer;
-  background-color: #274e13;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  font-size: 14px;
-}
-
-.save-changes:hover,
-.edit-changes:hover,
-.delete-item-btn:hover {
-  cursor: pointer;
-  transition: transform 0.3s ease-in-out;
-  transform: scale(1.05);
-}
-
 .edit-changes {
   background-color: #ce7e00;
 }
 .vendor-container {
   overflow: scroll;
-}
-
-.delete-item-btn {
-  padding: 4px 8px;
-  cursor: pointer;
-  background-color: #274e13;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  font-size: 14px;
-}
-
-.item-btns {
-  display: flex;
-  gap: 1rem;
-  align-items: center;
 }
 
 @media (max-width: 1077px) {
@@ -186,13 +116,5 @@
   }
   .post-btn {
     font-size: 1.2rem;
-  }
-}
-
-@media (max-width: 711px) {
-  .item-details {
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: flex-start;
   }
 }
